CM-1633: Moved ArtifactAsset into artifact package for better clarity.
