+++ comet-java-client/src/main/java/ml/comet/experiment/artifact/Artifact.java
@@ -2,7 +2,6 @@ package ml.comet.experiment.artifact;
 
 import lombok.NonNull;
 import ml.comet.experiment.impl.ArtifactImpl;
-import ml.comet.experiment.asset.ArtifactAsset;
 
 import java.io.File;
 import java.io.IOException;
