+++ comet-java-client/src/main/java/ml/comet/experiment/impl/ArtifactImpl.java
@@ -6,10 +6,9 @@ import lombok.NonNull;
 import ml.comet.experiment.artifact.Artifact;
 import ml.comet.experiment.artifact.ArtifactBuilder;
 import ml.comet.experiment.artifact.ConflictingArtifactAssetNameException;
-import ml.comet.experiment.asset.ArtifactAsset;
+import ml.comet.experiment.artifact.ArtifactAsset;
 import ml.comet.experiment.impl.asset.ArtifactAssetImpl;
 import ml.comet.experiment.impl.asset.ArtifactRemoteAssetImpl;
-import ml.comet.experiment.asset.Asset;
 import ml.comet.experiment.asset.RemoteAsset;
 import ml.comet.experiment.impl.asset.AssetImpl;
 import org.slf4j.Logger;
